Move type-specific generator model code outside the main generator module.
